The project ideas concerning the software
This software will help the company to be more efficient in handling the daily activities and registration of the patients. The purpose of this project is to give a complete requirement documentations, design, and implementation of the software. It also explains the user interface, hardware and software and different models that could be used to develop software.